First working version. Not perfect, but it works.
I had to require FastLED to be compatible with CRGB, not sure if
RGB88 can be made to work.
SMshow copy should be in the library, maybe moved as a #define macro

All in all it's a terrible hack, but I have no better idea given the
heavy limitations coming from SmartMatrix templates.