Make show() a callback so that it can be passed to the object.
Now matrix->show() works again.

This is not an ideal solution given that a copy function must be defined
by the caller, but it seems like the least crappy solution since at
least the rest of the code can call ->show() as usual and not worry
about the different implementation in SmartMatrix.