Import larger matrix improvements from Jason Coon and my own
https://github.com/pixelmatix/aurora/commit/8df7706ea41ba228870dd97e98073662c7e13c52
and
https://github.com/pixelmatix/aurora/commit/7b78ce3eb08b73ab19748fb9ed27de451e61b48d

Also added my own changes and improvements during testing on 64x96
display.
I made my predators not avoid borders as bouncing off doesn't look good
as it's a clip and not a bounce.

Notes:
- flowfield looks much cooler now that it can flow sideways
- incrementaldrift 1 and 2 look very different on a non square display
  and can be tweaked to give yet more cool effects.
- Swirl still looks bad on a big display, the dots are too small. I
  could fit it, but I'm not a huge fan of that effect anyway.

I checked some of the fixed patterns under notmytaste and they're still
not that great for me.