Big layer of my fixes on top of Mark's pristine tree.
1) arduino compiler allows to use a function that hasn't been defined
yet. That's illegal in C, so this doesn't work if you don't use the
arduino compiler.
The fix is just to put a bunch of function headers in Table_Mark_Estes_Impl.h

2) the_big_dog.h.ino -> Table_Mark_Estes_Impl.h
for the same problem with includes I mentioned, the_big_dog.h.ino is not
legal to copmile on its own (normal outside compilers will try to do it and
fail). All I did was rename it to .h so that it doen't get picked up as
something that needs to be compiled on its own.

3) TME_AUDIO / EasyTransfer. Your audio stuff does not build outside of
your teensy setup. I made an ifdef that is on by default for you, so it
will just keep working for you, but will be disabled by default on rPi.

4) Drawable => AuroraDrawable I had to rename this because it conflict
with another X11 Drawable object when building on linux

5) byte => int16_t
I sent you that one earlier, this is required for anything past 255.
I had to refix them all. If you could keep them, that will help you when
you go to bigger sizes and save me the work of re-patching :)

6) a bunch of other fixes for sizes bigger than 256
I changed the obvious beatsin8 into beatsin16. I didn't chagne them all,
just the ones that needed it according to -Woverflow from the compiler

7) I also had to fix Aurora which wasn't designed for more than 255x255

8) break in the wrong place or missing (sublte bugs, or at least compiler warnings)
case 22:
case 55:
case 128:
case 142