Switch to using mmin/mmax and more data size length changes.
min/max are unsafe on arduino, they mean different things on
different architectures (often templates that cannot compare different
types).
Define mmin/mmax instead to not conflict with the stock ones, and use
those instead.

Got rid of all byte and replaced with uint16_t. It might be bigger
than needed for some types, but it was easier than checking them
all one by one (which I did for arrays).

Disabled fcool and heaty as they are not used (Fire disabled).

Tested on ESP32 and linux in 384x256 resolution. This fixes the
problems and hangs I saw before, as well as takes care of all
the compiler warnigns I got.